refactor(epicenter): consolidate schema converters into converters/ folder
Move all schema-to-X conversion utilities (arktype, drizzle, json-schema) into
a dedicated converters/ folder. This prepares for removing ~standard from field
schemas and using converters as the single source for validation.
