+++ packages/epicenter/src/providers/sqlite/schema/index.ts
@@ -4,12 +4,12 @@
  * Column builders and schema conversion tools for working with Drizzle ORM.
  */
 
-export type { WorkspaceSchemaToDrizzleTables } from '../../../core/schema/fields/to-drizzle';
+export type { WorkspaceSchemaToDrizzleTables } from '../../../core/schema/converters/to-drizzle';
 
 // Schema converter (Epicenter → Drizzle)
 export {
 	convertTableSchemaToDrizzle,
 	convertWorkspaceSchemaToDrizzle,
-} from '../../../core/schema/fields/to-drizzle';
+} from '../../../core/schema/converters/to-drizzle';
 // Column builders for defining Drizzle schemas
 export { boolean, date, id, integer, real, tags, text } from './builders';
